+++ CHANGELOG.md
@@ -61,6 +61,7 @@ Extensions
 - `URLRequest.addHTTPHeaderFields(_:)`, `URLRequest.addHTTPHeaderFields(json:)`, and `URLRequest.addHTTPHeaderFields(object:)` methods are added, that add values to header fields
 - `URL.init(string:pathParameters:queryParameters:)` method is added, that initializes `URL` with `String`, path parameters, and query parameters
 - `CGSize.scaledUp(withConstant:)`, `CGSize.scaledDown(withConstant:)`, `CGSize.scaledUp(withMultiplier:)`, and `CGSize.scaledDown(withMultiplier:)` methods, alongside their mutator counterparts, are added for scaling `CGSize`
+-  `CGPoint.movedLeft(withValue:)`, `CGPoint.movedRight(withValue:)`, `CGPoint.movedUp(withValue:)`, and `CGPoint.movedDown(withValue:)` methods, alongside their mutator counterparts, are added for scaling `CGSize`
 
 API